Allow custom text formatting
Reading an optional `format_text` function to treat each segment of the
test.

For example, when working with test suites, the navigation looks like

`describe('something') callback > describe ('subcategory') callback > it('can do things') callback`

The text is useful, but it can be a bit too verbose.

It would be great if we could shape the segments to be something like
this:

`describe something > describe  subcategory > it can do things`

Which can be achieved with something like:

```fennel
(navic.setup {:format_text (fn [text]
                             (if (or (text:match "^it%(")
                                     (text:match "^describe%("))
                               (-> text
                                 (string.gsub "^it%('" "it ")
                                 (string.gsub "^describe%('" "describe ")
                                 (string.gsub "'%) callback$" ""))
                               text))
```